+++ src/biz/bokhorst/xprivacy/Util.java
@@ -34,18 +34,12 @@ import android.util.Log;
 
 public class Util {
 	private static boolean mPro = false;
-	private static boolean mLog = true;
-	private static boolean mLogDetermined = false;
 
 	public static void log(XHook hook, int priority, String msg) {
-		// Check if logging enabled
-		if (Process.myUid() != 0 && !mLogDetermined) {
-			mLogDetermined = true;
-			mLog = PrivacyManager.getSettingBool(null, null, 0, PrivacyManager.cSettingLog, false, false);
-		}
-
 		// Log if enabled
-		if (priority != Log.DEBUG && (priority == Log.INFO ? mLog : true))
+		boolean log = (Process.myUid() == 0 || PrivacyManager.getSettingBool(null, null, 0, PrivacyManager.cSettingLog,
+				false, false));
+		if (priority != Log.DEBUG && (priority == Log.INFO ? log : true))
 			if (hook == null)
 				Log.println(priority, "XPrivacy", msg);
 			else